requirements.txt
# Django et REST framework
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
django-filter==23.3

# Base de données
psycopg2-binary==2.9.10
dj-database-url==2.1.0

# Gestion des variables d'environnement
python-decouple==3.8

# Traitement des données
pandas==2.2.3
openpyxl==3.1.2
numpy==2.3.2
python-dateutil==2.9.0
pytz==2025.1
xlswriter==3.2.5

# Utilitaires
Pillow==11.2.1
django-q==1.3.9

# Serveur WSGI
gunicorn==21.2.0
whitenoise==6.5.0